backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.58.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
redis==5.2.1
pydantic==2.10.3
python-multipart==0.0.20
httpx==0.28.1 